Move \AdvancedCustomBlocks\View\BlockPostType to \AdvancedCustomBlocks\PostTypes\BlockPostType
